Actually persist NFTs
