+++ packages/shared/src/lib/core/nfts/actions/checkForUntrackedNfts.ts
@@ -4,14 +4,11 @@ import { EvmExplorerApi } from '@core/network/classes'
 import { getNetwork } from '@core/network/stores'
 import { TokenTrackingStatus } from '@core/token/enums'
 
+import { DEFAULT_NFT_TOKEN_ID } from '../constants'
 import { NftStandard } from '../enums'
-import {
-    IErc721ContractMetadata,
-    IErc721TokenMetadata,
-    IErc721TokenMetadataAttribute,
-    IPersistedErc721Nft,
-} from '../interfaces'
-import { composeUrlFromNftUri } from '../utils'
+import { IErc721ContractMetadata } from '../interfaces'
+import { addPersistedNft } from '../stores'
+import { getPersistedErc721NftFromContract } from '../utils'
 import { addNewTrackedNftToActiveProfile } from './addNewTrackedNftToActiveProfile'
 
 export function checkForUntrackedNfts(account: IAccountState): void {
@@ -29,9 +26,7 @@ export function checkForUntrackedNfts(account: IAccountState): void {
         for (const explorerNft of explorerNfts) {
             const { token, value } = explorerNft
             const { address, name, symbol } = token
-
-            addNewTrackedNftToActiveProfile(networkId, address, TokenTrackingStatus.AutomaticallyTracked)
-
+            const contract = chain.getContract(ContractType.Erc721, address)
             const contractMetadata: IErc721ContractMetadata = {
                 standard: NftStandard.Erc721,
                 address,
@@ -39,65 +34,27 @@ export function checkForUntrackedNfts(account: IAccountState): void {
                 symbol,
             }
 
-            const contract = chain.getContract(ContractType.Erc721, address)
             const isEnumerable = await contract.methods.supportsInterface(Erc721InterfaceId.Enumerable).call()
-
             if (isEnumerable) {
-                for (let idx = 0; idx < Number(value); idx++) {
-                    const tokenId = await contract.methods.tokenOfOwnerByIndex(evmAddress, idx).call()
-                    const persistedNft: IPersistedErc721Nft = {
-                        standard: NftStandard.Erc721,
-                        contractMetadata,
-                        tokenId,
-                    }
-
-                    const hasTokenMetadata = await contract.methods.supportsInterface(Erc721InterfaceId.Metadata).call()
-                    if (hasTokenMetadata) {
-                        const tokenUri = await contract.methods.tokenURI(tokenId).call()
-                        const composedTokenUri = composeUrlFromNftUri(tokenUri)
-                        persistedNft.tokenUri = composedTokenUri
-
-                        const response = await fetch(composedTokenUri)
-                        const metadata = (await response.json()) as IErc721TokenMetadata
-                        if (metadata) {
-                            const attributes: IErc721TokenMetadataAttribute[] = metadata.attributes?.map(
-                                (attribute) => ({ traitType: attribute['trait_type'], value: attribute.value })
-                            )
-                            persistedNft.tokenMetadata = {
-                                ...metadata,
-                                image: composeUrlFromNftUri(metadata.image) ?? metadata.image,
-                                attributes,
-                            }
-                        }
-                    }
-                }
+                await Promise.all(
+                    Array.from({ length: Number(value) }).map(async (_, idx) => {
+                        const tokenId = await contract.methods.tokenOfOwnerByIndex(evmAddress, idx).call()
+                        // only if NOT already persisted
+                        addPersistedNft(
+                            `${address}:${tokenId}`,
+                            await getPersistedErc721NftFromContract(tokenId, contract, contractMetadata)
+                        )
+                    })
+                )
             } else {
-                const tokenId = '0'
-                const persistedNft: IPersistedErc721Nft = {
-                    standard: NftStandard.Erc721,
-                    contractMetadata,
-                    tokenId,
-                }
-
-                const hasTokenMetadata = await contract.methods.supportsInterface(Erc721InterfaceId.Metadata).call()
-                if (hasTokenMetadata) {
-                    const tokenUri = await contract.methods.tokenURI('0').call()
-                    persistedNft.tokenUri = tokenUri
-
-                    const response = await fetch(tokenUri)
-                    const metadata = (await response.json()) as IErc721TokenMetadata
-                    if (metadata) {
-                        const attributes: IErc721TokenMetadataAttribute[] = metadata.attributes?.map((attribute) => ({
-                            traitType: attribute['trait_type'],
-                            value: attribute.value,
-                        }))
-                        persistedNft.tokenMetadata = {
-                            ...metadata,
-                            attributes,
-                        }
-                    }
-                }
+                // only if NOT already persisted
+                addPersistedNft(
+                    `${address}:${DEFAULT_NFT_TOKEN_ID}`,
+                    await getPersistedErc721NftFromContract(DEFAULT_NFT_TOKEN_ID, contract, contractMetadata)
+                )
             }
+
+            addNewTrackedNftToActiveProfile(networkId, address, TokenTrackingStatus.AutomaticallyTracked)
         }
     })
 }
