+++ packages/shared/src/lib/core/nfts/constants/default-nft-token-id.constant.ts
@@ -1,0 +1,1 @@
+export const DEFAULT_NFT_TOKEN_ID: string = '0'
