+++ packages/shared/src/lib/core/nfts/utils/index.ts
@@ -2,6 +2,7 @@ export * from './checkIfNftShouldBeDownloaded'
 export * from './composeUrlFromNftUri'
 export * from './convertAndFormatNftMetadata'
 export * from './getNftsFromNftIds'
+export * from './getPersistedErc721NftFromContract'
 export * from './getSpendableStatusFromUnspentNftOutput'
 export * from './fetchWithTimeout'
 export * from './isNftOwnedByAnyAccount'
